# api/swagger.json
{"openapi":"3.0.1","info":{"title":"Spring auth backend service","description":"Сервис авторизации и управления пользователями для мобильного приложения","contact":{"name":"Evgeny Gribanov","url":"https://codecow.pw","email":"gribanoveu@yandex.com"},"version":"1.0.0"},"servers":[{"url":"https://localhost:8443/api","description":"Generated server url"}],"tags":[{"name":"Аутентификация пользователя","description":"Позволяет получить и обновить токен"},{"name":"Контроллер для модерации","description":"Управляет ограничениями пользователя"},{"name":"Контроллер пользователя","description":"Юзер управляет своим аккаунтом"},{"name":"Контроллер для назначения модераторов","description":"Управляет ролями пользователя"},{"name":"Управление аккаунтом без авторизации","description":"Позволяет управлять данными учетной записи"},{"name":"Управление аккаунтом с авторизацией","description":"Позволяет управлять данными учетной записи"}],"paths":{"/open/v1/auth":{"post":{"tags":["Аутентификация пользователя"],"summary":"Получить токен","description":"Получение access token и refresh token","operationId":"authenticateUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}}}},"patch":{"tags":["Аутентификация пользователя"],"summary":"Обновить токен","description":"Получение нового токена на основе refresh token","operationId":"refreshToken","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshTokenDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}}}}},"/open/v1/account/restore-password":{"post":{"tags":["Управление аккаунтом без авторизации"],"summary":"Восстановить пароль через OTP код","operationId":"restorePassword","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RestorePasswordDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StatusResponse"}}}}}}},"/open/v1/account/generate-code":{"post":{"tags":["Управление аккаунтом без авторизации"],"summary":"Сгенерировать OTP код","operationId":"generateOtpCode","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenerateOtpDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StatusResponse"}}}}}}},"/open/v1/account/create":{"post":{"tags":["Управление аккаунтом без авторизации"],"summary":"Регистрация нового пользователя","operationId":"registerUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StatusResponse"}}}}}}},"/v1/role":{"get":{"tags":["Контроллер для назначения модераторов"],"summary":"Получить роль пользователя","operationId":"getUserRole","parameters":[{"name":"email","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StatusResponse"}}}}},"security":[{"JWT":[]}]},"delete":{"tags":["Контроллер для назначения модераторов"],"summary":"Обновить роль до пользователя","operationId":"updateToUser","parameters":[{"name":"email","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StatusResponse"}}}}},"security":[{"JWT":[]}]},"patch":{"tags":["Контроллер для назначения модераторов"],"summary":"Обновить роль до модератора","operationId":"updateToModerator","parameters":[{"name":"email","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StatusResponse"}}}}},"security":[{"JWT":[]}]}},"/v1/moderator/mercy":{"patch":{"tags":["Контроллер для модерации"],"summary":"Снять бан с пользователя","operationId":"mercyUser","parameters":[{"name":"email","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StatusResponse"}}}}},"security":[{"JWT":[]}]}},"/v1/moderator/enable":{"patch":{"tags":["Контроллер для модерации"],"summary":"Включить пользователя","operationId":"enableUser","parameters":[{"name":"email","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StatusResponse"}}}}},"security":[{"JWT":[]}]}},"/v1/moderator/disable":{"patch":{"tags":["Контроллер для модерации"],"summary":"Отключить пользователя","operationId":"disableUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RestrictionDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StatusResponse"}}}}},"security":[{"JWT":[]}]}},"/v1/moderator/ban":{"patch":{"tags":["Контроллер для модерации"],"summary":"Выдать бан пользователю","operationId":"banUserReq","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RestrictionDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StatusResponse"}}}}},"security":[{"JWT":[]}]}},"/v1/account/change-password":{"patch":{"tags":["Управление аккаунтом с авторизацией"],"summary":"Сменить пароль","operationId":"changePassword","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePasswordDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StatusResponse"}}}}},"security":[{"JWT":[]}]}},"/v1/account/change-email":{"patch":{"tags":["Управление аккаунтом с авторизацией"],"summary":"Сменить Email","operationId":"changeEmail","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangeEmailDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StatusResponse"}}}}},"security":[{"JWT":[]}]}},"/v1/user":{"get":{"tags":["Контроллер пользователя"],"summary":"Получить данные текущего пользователя","operationId":"getUserData","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}},"security":[{"JWT":[]}]},"delete":{"tags":["Контроллер пользователя"],"summary":"Удалить текущего авторизованного пользователя","operationId":"deleteUser","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StatusResponse"}}}}},"security":[{"JWT":[]}]}},"/v1/role/users":{"get":{"tags":["Контроллер для назначения модераторов"],"summary":"Получить список всех пользователей","operationId":"getUsersList","parameters":[{"name":"page","in":"query","description":"Zero-based page index (0..N)","required":false,"schema":{"minimum":0,"type":"integer","default":0}},{"name":"size","in":"query","description":"The size of the page to be returned","required":false,"schema":{"minimum":1,"type":"integer","default":10}},{"name":"sort","in":"query","description":"Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.","required":false,"schema":{"type":"array","items":{"type":"string"}}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UsersResponse"}}}}},"security":[{"JWT":[]}]}},"/v1/role/moders":{"get":{"tags":["Контроллер для назначения модераторов"],"summary":"Получить список всех модераторов","operationId":"getModersList","parameters":[{"name":"page","in":"query","description":"Zero-based page index (0..N)","required":false,"schema":{"minimum":0,"type":"integer","default":0}},{"name":"size","in":"query","description":"The size of the page to be returned","required":false,"schema":{"minimum":1,"type":"integer","default":10}},{"name":"sort","in":"query","description":"Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.","required":false,"schema":{"type":"array","items":{"type":"string"}}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UsersResponse"}}}}},"security":[{"JWT":[]}]}},"/v1/moderator/users":{"get":{"tags":["Контроллер для модерации"],"summary":"Получить список всех пользователей","operationId":"getAllUsersList","parameters":[{"name":"page","in":"query","description":"Zero-based page index (0..N)","required":false,"schema":{"minimum":0,"type":"integer","default":0}},{"name":"size","in":"query","description":"The size of the page to be returned","required":false,"schema":{"minimum":1,"type":"integer","default":10}},{"name":"sort","in":"query","description":"Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.","required":false,"schema":{"type":"array","items":{"type":"string"}}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}},"security":[{"JWT":[]}]}},"/open/v1/account/restrictions":{"get":{"tags":["Управление аккаунтом без авторизации"],"summary":"Получить причину блокировки","operationId":"getRestrictionsReason","parameters":[{"name":"email","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StatusResponse"}}}}}}},"/v1/moderator/delete":{"delete":{"tags":["Контроллер для модерации"],"summary":"Удалить пользователя","operationId":"deleteUser_1","parameters":[{"name":"email","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StatusResponse"}}}}},"security":[{"JWT":[]}]}}},"components":{"schemas":{"LoginDto":{"required":["email","password"],"type":"object","properties":{"email":{"maxLength":80,"minLength":0,"pattern":"^[_A-Za-z0-9-+]+(\\.[_A-Za-z0-9-]+)*@[A-Za-z0-9-]+(\\.[A-Za-z0-9]+)*(\\.[A-Za-z]{2,})$","type":"string","description":"Электронная почта","example":"user@email.com"},"password":{"maxLength":80,"minLength":0,"type":"string","description":"Пароль пользователя","example":"Qwerty123"}},"description":"Сущность для авторизации"},"TokenResponse":{"type":"object","properties":{"timestamp":{"type":"string","description":"Время выполнения запроса","format":"date-time"},"access_token":{"type":"string","description":"Токен доступа"},"token_type":{"type":"string","description":"Тип токена","default":"Bearer"},"expires_in":{"type":"integer","description":"Время жизни токена в секундах","format":"int64","default":900},"refresh_token":{"type":"string","description":"Токен для обновления"}},"description":"Ответ после успешной авторизации"},"RestorePasswordDto":{"required":["confirmPassword","email","otpCode","password"],"type":"object","properties":{"email":{"maxLength":30,"minLength":0,"pattern":"^[_A-Za-z0-9-+]+(\\.[_A-Za-z0-9-]+)*@[A-Za-z0-9-]+(\\.[A-Za-z0-9]+)*(\\.[A-Za-z]{2,})$","type":"string","description":"Электронная почта","example":"user@email.com"},"otpCode":{"pattern":"^\\d{6}$","type":"string","description":"Временный код подтверждения OTP"},"password":{"maxLength":80,"minLength":0,"type":"string","description":"Пароль"},"confirmPassword":{"maxLength":80,"minLength":0,"type":"string","description":"Пароль повторно"}},"description":"Сущность для восстановления пароля"},"ResponseDetails":{"type":"object","properties":{"type":{"type":"string","description":"Тип операции"},"code":{"type":"string","description":"Код операции"},"message":{"type":"string","description":"Подробное описание операции"}},"description":"Сущность для описания деталей операции"},"StatusResponse":{"type":"object","properties":{"timestamp":{"type":"string","description":"Время выполнения запроса","format":"date-time"},"status":{"type":"string","description":"Уровень статуса запроса","default":"INFO","enum":["INFO","WARNING","ERROR","SUCCESS"]},"details":{"type":"array","description":"Детали запроса","items":{"$ref":"#/components/schemas/ResponseDetails"}}},"description":"Сущность для возврата статуса операции"},"GenerateOtpDto":{"required":["email"],"type":"object","properties":{"email":{"maxLength":30,"minLength":0,"pattern":"^[_A-Za-z0-9-+]+(\\.[_A-Za-z0-9-]+)*@[A-Za-z0-9-]+(\\.[A-Za-z0-9]+)*(\\.[A-Za-z]{2,})$","type":"string","description":"Email","example":"user@email.com"}},"description":"Сущность для генерации временного кода OTP"},"RegisterDto":{"required":["confirmPassword","email","password"],"type":"object","properties":{"email":{"maxLength":80,"minLength":0,"pattern":"^[_A-Za-z0-9-+]+(\\.[_A-Za-z0-9-]+)*@[A-Za-z0-9-]+(\\.[A-Za-z0-9]+)*(\\.[A-Za-z]{2,})$","type":"string","description":"Электронная почта","example":"user@email.com"},"birthDate":{"type":"string","description":"Дата рождения","example":"2001-12-13"},"password":{"maxLength":80,"minLength":0,"type":"string","description":"Пароль"},"confirmPassword":{"maxLength":80,"minLength":0,"type":"string","description":"Пароль повторно"}},"description":"Сущность для регистрации нового пользователя"},"RestrictionDto":{"required":["reasonCode","userEmail"],"type":"object","properties":{"userEmail":{"maxLength":30,"minLength":0,"pattern":"^[_A-Za-z0-9-+]+(\\.[_A-Za-z0-9-]+)*@[A-Za-z0-9-]+(\\.[A-Za-z0-9]+)*(\\.[A-Za-z]{2,})$","type":"string","description":"Email","example":"user@email.com"},"reasonCode":{"maxLength":15,"minLength":0,"type":"string","description":"Пункт правил причины","example":"CR-1.1"}},"description":"Сущность для установки ограничений на пользователя"},"ChangePasswordDto":{"required":["confirmPassword","oldPassword","password"],"type":"object","properties":{"oldPassword":{"maxLength":80,"minLength":0,"type":"string","description":"Старый пароль","example":"Qwerty123"},"password":{"maxLength":80,"minLength":0,"type":"string","description":"Новый пароль","example":"Qwerty1234"},"confirmPassword":{"maxLength":80,"minLength":0,"type":"string","description":"Повторный ввод нового пароля","example":"Qwerty1234"}},"description":"Сущность для смены пароля авторизованным пользователем"},"ChangeEmailDto":{"required":["email"],"type":"object","properties":{"email":{"maxLength":30,"minLength":0,"pattern":"^[_A-Za-z0-9-+]+(\\.[_A-Za-z0-9-]+)*@[A-Za-z0-9-]+(\\.[A-Za-z0-9]+)*(\\.[A-Za-z]{2,})$","type":"string","description":"Электронная почта","example":"user@email.com"}},"description":"Сущность для смены email авторизованным пользователем"},"RefreshTokenDto":{"required":["refreshToken"],"type":"object","properties":{"refreshToken":{"type":"string","description":"Токен обновления (refresh token)"}},"description":"Сущность для получения нового access token"},"User":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"version":{"type":"integer","format":"int64"},"email":{"type":"string"},"password":{"type":"string","writeOnly":true},"registrationDate":{"type":"string","format":"date-time"},"birthDate":{"type":"string","format":"date"},"banExpiration":{"type":"string","format":"date-time"},"restrictionReason":{"type":"string"},"accountNonExpired":{"type":"boolean"},"accountNonLocked":{"type":"boolean"},"credentialsNonExpired":{"type":"boolean"},"enabled":{"type":"boolean"}},"description":"Базовая сущность пользователя"},"UsersResponse":{"type":"object","properties":{"timestamp":{"type":"string","description":"Время выполнения запроса","format":"date-time"},"status":{"type":"string","description":"Уровень статуса запроса","default":"INFO","enum":["INFO","WARNING","ERROR","SUCCESS"]},"users":{"type":"array","description":"Информация о пользователе","items":{"$ref":"#/components/schemas/User"}}},"description":"Ответ со списком всех пользователей"}},"securitySchemes":{"JWT":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}